pyproject.toml
[project]
name = "density"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "David Ciani", email = "dciani@davidciani.com" }
]
requires-python = ">=3.13"
dependencies = [
    "geoarrow-pyarrow>=0.2.0",
    "geopandas>=1.1.1",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "osmnx>=2.0.5",
    "pandas>=2.3.1",
    "pyarrow>=20.0.0",
    "pyarrow-stubs>=20.0.0.20250716",
    "pyproj>=3.7.1",
    "quackosm>=0.15.2",
    "rasterio>=1.4.3",
    "rich>=14.0.0",
    "shapely>=2.1.1",
]

[project.scripts]
density = "density:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
